Using ffmpeg for everything
