+++ src/subtitle_tool/audio.py
@@ -1,6 +1,8 @@
+import io
 import logging
 from dataclasses import dataclass
 
+import ffmpeg
 from humanize.time import precisedelta
 from pydub import AudioSegment, silence
 
@@ -112,3 +114,60 @@ class AudioSplitter:
         )
 
         return segments
+
+
+class AudioExtractionError(Exception):
+    pass
+
+
+def extract_audio(file_path: str) -> AudioSegment:
+    """
+    Extract an audio stream from any file recognized by ffmpeg.
+    This method will always extract the main audio stream.
+    The audio stream will be extracted to wav as it's the closest from
+    the PCM format that pydubs uses internally.
+
+    Args:
+        file_path (str): path to the media file
+
+    Returns:
+        AudioSegment: in-memory representation of the extracted audio stream.
+    """
+    if not file_path:
+        raise AudioExtractionError("Path to media file is mandatory")
+
+    try:
+        probe = ffmpeg.probe(file_path)
+    except ffmpeg.Error as e:
+        logger.error(f"Error probing for file metadata: {e}")
+        raise AudioExtractionError("Error probing for file metadata") from e
+
+    audio_streams = [s for s in probe["streams"] if s["codec_type"] == "audio"]
+    if not audio_streams:
+        raise AudioExtractionError(f"No audio streams found in file {file_path}")
+
+    audio_stream = audio_streams[0]
+    audio_codec = audio_stream.get("codec_name", "")
+    logger.info(f"Audio stream detected: {audio_codec}")
+
+    # Extract audio
+    logger.debug(f"Converting {audio_codec} to wav")
+    audio_format = "wav"
+    process = (
+        ffmpeg.input(file_path)
+        .output("pipe:", format=audio_format, acodec="pcm_s16le")
+        .run_async(pipe_stdout=True, pipe_stderr=True)
+    )
+
+    out, err = process.communicate()
+
+    # Check the return code instead of stderr content
+    if process.returncode != 0:
+        logger.error(f"Extraction error with ffmpeg: {err.decode()}")
+        raise AudioExtractionError(f"Extraction error: {err.decode()}")
+
+    audio_buffer = io.BytesIO(out)
+
+    # Convert audio to AudioSegment
+    audio_buffer.seek(0)
+    return AudioSegment.from_file(audio_buffer, format=audio_format)
